// app/src/app/pokemon/pokemon.interfaces.ts
export interface ICries {
    latest: string
    legacy: string
}

export interface ISprites {
    back_default: string
    back_female: string
    back_shiny: string
    back_shiny_female: string
    front_default: string
    front_female: string
    front_shiny: string
    front_shiny_female: string
}

export interface IMove {
    name: string
    url: string
}

export interface IAbility {
    name: string
    url: string
}

export interface IPokemonItem {
    id: string
    name: string
    // url: string
}

export interface IPokemon extends Omit<IPokemonItem, 'url'> {
    cries: ICries
    sprites: ISprites
    weight: number
    height: number
    moves: IMove[]
    ability: IAbility[]
    descripton?: string

}

export interface IResult {
    count: number
    next: string | null
    previous: string | null
    results: IPokemonItem[]
}